undo rename
